fix: Apply theme borderRadius to ExpansionPanel
`ExpansionPanel` has a new prop `square` defaulted to false.
`rounded` style is applied otherwise. By default, borderRadius
is only applied to firstChild and lastChild respectively on
top and bottom. With this change, all corners of our
`ExpansionPanel`s are rounded.

One caveat: if someone tries to customize the theme, they will
need to set the borderRadius again on `ExpansionPanel` (it will
not suffice to change the borderRadius of `theme.shape`, I haven\'t
found a way to access the shape of the *current* theme while
overriding the borderRadius of `ExpansionPanel`.