feat(utils): Get CSS variable value from body
Stack themes are overriding CSS variable on body. So getting it from
html can lead to incorrect values when using a theme from the stack.

For example, in an app that uses a stack theme, I can be in the
following situation:

```
:root {
  --primaryColor: black;
}

body {
  --primaryColor: white;
}
```

In that case, if we get the variable value from `html`, we will get
`black`, where we want to get `white`. With this fix, we will get
`white`.