fix: Use browserslist-config-cozy instead of custom autoprefixer config
In postcss config, we gave the browsers list to autoprefixer plugin
directly. But the `browsers` property is deprecated and will be removed
in next major version (see
https://github.com/postcss/autoprefixer/releases/tag/9.6.0).

We also use autoprefixer through autoprefixer-stylus, but it uses a very
old version of autoprefixer (7.1.0) that doesn't understand the `dead`
query that our browserslist config uses. A PR fixing this have been
merged and will hopefuly be released soon, see
https://github.com/jescalan/autoprefixer-stylus/pull/180. That's why I
left the custom browserslist config for autoprefixer-stylus in
`build:css`, `build:css:utils` and `build:css:kss` scripts for now. But
we should upgrade this module and stop using a custom browserslist
config for it as soon as possible.