fix(ActionMenuItem): Handle onClick on Item
Before this commit, if we attached an onClick to an
ActionMenuItem, the onClick was not called since
the propagation of the event was not stopped
resulting in the close of the menu instead of the
call of the onClick props.

With this fix, we stop the propagation if an onClick
is here, and we call it after.