fix: Stack spacing xs and s work
Since the class for `m` spacing was outputted in the CSS after the class
for xs and s, it "won" over the xs. Solution is to remove the class "m"
that was by default on the component (under the name .Stack) and to only
apply to the component the spacing class.