feat(PercentageLine): Make it customizable
In the context of consumer credit details page in banks, we need to show
a percentage line that is slightly different from the one we have in UI.
All customizations (like adding a background-color, a track border) were
not possible with the current implementation. With these changes, we can
customize it more.