fix: Always use transpiled cozy-ui (#1307)
Some files did use an absolute path for cozy-ui
loading from `cozy-ui/react/…`. This breaks
if we use the components from a transpiled path
(it will escape the transpiled to go into non transpiled)

----

In early versions of B2B overrides, we needed to have absolute path for the files we wanted to overide. Files modified in this PR were using absolute path for that motive.

Today, our babel configuration does rewrite all path in order to use absolute path pointing at the transpiled version of each file. We don't need to write absolute path ourself in cozy-ui anymore (and we shouldn't).