feat(TextField): Add inverted prop
This prop is used to show an « inverted » version of the TextField. E.g.
a version that looks good on a primary color background.

MUI doesn't have a prop to manage different « looks » of the TextField,
so I cheated by wrapping the component in a `div` with a particular
class name when `inverted` is true.