feat(ContactsListModal): Double realtime with mobile app resume events
When using a mobile app, when the user wants to add a contact, the app
will be put in the background while the user is in a browser to add the
contact. During this time, the realtime events are not handled by the
app. To make sure the user sees fresh data, we have to also subscribe to
the cordova's resume event to re-fetch it when the app is put to
foreground again.