feat: Allows undefined values in useEventListener
To attach an event to an element, we often use a ref
to this element. The ref begins with a null value
which breaks useEventListener.

This change allows to keep the call to the hook
whatever the argument are, and only attach an
event if we have something to work with.