fix(PasswordInput): The getStrength function was not pure
Due to how RegExp.test works with RegExp with the `g` flag, the
getStrength function returned a different result when called multiple
times with the same inputs.
See https://frama.link/d5dgGURD for more infos.

This problem appeared when we extracted the regexp from the function.
Before, they were recreated on each call, which was more expensive, but
not buggy. To fix the problem without putting the regexp in the function
again, we can just remove the `g` flags that are useless here.