feat: Optimize the output of the CSS (#1379)
The minified CSS is quite heavy, and a large part of the wright is due
to media queries. I think we can do a lot better. The first optimization
is to regroup media queries from the cssModulesUtils and nativeUtils
mixins. Before this commit, cozy-ui.min.css took 129kb, and after, it
tooks now 123kb.