fix: Popover position
Before, the popover was portaled. It worked well
in a normal context, but not when we only have
a portal like a Modal.

The idea, is to display the popover in the
component calling it and put it in a fixed
position (we can't have an absolute one
since Modals for instance are fixed...).

We set the right zIndex to this popover to be well
displayed.

The drawback of this method, is that the Popover
doesn't follow the scroll anymore. But we already
have this behavior in our old Menu. I think we can
deal with it, the time we find a solution.