feat: Make full width work with the stack
The full-width class no longer sets all margins, but only
the horizontal ones, which makes it possible to use with
the Stack component.

Fixes #1561

BREAKING CHANGE: Button in full width not longer resets the vertical
margin. This can break situation when the Button is within a parent
imposing a vertical margin for its children, like Stack for example.
Before, the Button would not have any margin, but now it would follow
the parent margins.