feat: Use typography components for nav list
primaryText and primaryTextClassName are still supported (same for
secondary).

Also tweaked margins to correspond to design specs

To support the ellipsis, we use the padding right of the text element if
there is the secondaryAction. It works well in our cases where there is
a 24x24 icon on the right but it is brittle: the icon could be wider and
the ellipsis would be too far away.

Unfortunately we cannot do better here as the secondaryAction is in
position absolute; the text elements cannot take the icon into account
properly.