fix: Scroll Lock Issue
first: Opening an ActionMenu was not locking the scroll

second: We had an issue with body-scroll-lock and MUI's Modal
See https://github.com/mui-org/material-ui/issues/5750#issuecomment-471806500
for more informations. To resume: MUI adds overflow:hidden
on body when opening a Modal, body-scroll-lock too (throught Overlay
-> ActionMenu).

So when opening a Dialog from an ActionMenu, it's result on a conflict
between MUI and body-scroll-lock : We can't scroll anymore after
closing the dialog.

By using React-Remove-Scroll we're avoiding this issue.