fix: Use a precise selector for SelectionBar button hover
In https://github.com/cozy/cozy-ui/commit/9c543922553fe8d193ba6c38e3f0521f26c88e9e#r43258678 we wondered why the selector was on role=application. It was to be more precise than the style on the button and on the hover button of Cozy-UI (we have by default blue hover on a subtle button). Let's use a slightly more precise selector by manually specifying the hover to make sure we use the right color.