feat: Better alignment for Radio/Checkbox
The parent component of the radio/checkbox did not have the correct
bounding box which led to incorrect alignments. Here, the min-height and
min-width is fixed to the size of the radio/checkbox. An additional
gutter props is set to true by default and can be deactivated if we
want to cancel the default gutter.

BREAKING CHANGE: A margin-bottom has been removed from checkbox/radio,
you must check in your app that it has not broken anything