feat: Rework Dividers
BREAKING CHANGE: List, ListItem and Divider do not have the same
behavior: they do not add automatically a divider border; you have to
add a top Divider and use the divider prop on ListItems to achieve the
same behavior as before. This BC was necessary since we had put the
borders in our default theme and we could not achieve the "inset" effect
with default CSS borders, we needed an extra HTML element for that. We
are now closer to the original MUI implementation.

Summary of changes

```
- import { CardDivider } from 'cozy-ui/transpiled/react/Divider'
+ import Divider from 'cozy-ui/transpiled/react/Divider'

- import Divider from 'cozy-ui/transpiled/react/Divider'
+ import { DeprecatedDivider } from 'cozy-ui/transpiled/react/Divider'

- <List>
+ <Divider />
+ <List>

- <ListItem />
+ <ListItem divider />
```