fix: Fix closing issue with the cross on iOS
On iPhone (at least) we had a bug with few kinds
of Dialog:
- ConfirmDialog small

We can't close them by clicking on the top right cross.

In fact, this is when we have conflicts with few dom elements
 (when dialogContentInner withFluidActions).
The DialogCloseButton is "behing" others.

I don't know why it's work on Safari Dekstop but not on Safari Mobile...