fix: Prevent default when opening intent dialog
The problem was that a click on chip inside a row with an onClick
handler would trigger both the row and chip onClick handlers. Here,
I preventDefault() on the ev as a way of saying "this event is
already taken care of", and the click handler of the row checks
ev.defaultPrevented and does not do anything if it is set to true.

I thought of using stopPropagation() but this solution seems less
brutal.