feat: Ability to remove a notification programmatically
Alerter has a new method removeNotification that can be used to
programmatically remove a notification that has been shown.

State changes

Previously, there would be two states, one in the store, and one in
the Alerter. Now there is only the state of the store, and the Alerter
keeps a reference on it.

Id changes

Ids would be kept as the current length of the list but this could be
problematic since you could have a situation where two notifications
have the same id. Now we have an incremental id.