test: Fix sortered contacts
Since Node 11, sort() has changed passing from something
not stable to something stable.

We needed to change a few "edge cases" for instance:
When we sorted :
{givenName: "", familyName: "A"} and {giveName: "A", familyName: ""}
we compared A to ... A.
Previously the first one was put in first, now it's the second.

It's a pretty small change since the case will not occurs
frequently. Now the behavior is the same as the browser.