fix: Remove `height:auto` from `app-2panes-sticky` class
This class is used when the UI is split in two panes, usually when the
app has a sidebar and a main content

When using `height:auto`, this prevents the main content to be
scrollable if it is smaller than the navbar. Also the navbar is not
scrollable at all as its height won't adapt to the browser size

Another effect is that the main content is not scrollable to the end of
document when the navbar overflows the document as it virtually push
the document's limits outside of the browser view

This behavior is visible only in `cozy-store` as it is the only app
that have a navbar tall enough to overflow the browser's body

Other apps are not impacted as their navbar is very small

By removing `height:auto`, then the navbar height adapt to its
container size (vs its own size). This allow the main content to be
scrollable when smaller

This fix should be completed by some CSS edit on apps side with an
`overflow:auto` on the sidebar. But here again only `cozy-store` is
impacted for now