fix: Remove export Sprite from Icon (index.jsx)
We delete this export in order to optimize the size
of the Icon component.

As soon as we imported this Icon component we
embed Sprite which is not necessarily used.
This can increase the size of the
application that uses Cozy UI.

BREAKING CHANGE:
 `<Sprite>` is no longer exported by `<Icon>`. However, you can import it directly with `import Sprite from 'cozy-ui/transpiled/react/Icon/Sprite'`. Be carful, it's no longer possible to set an icon with a string like this `<Icon icon="myIcon" />` without importing `<Sprite>`. If don't need the sprite, you can still use svgr icons:
```
import MyIcon from 'cozy-ui/transpiled/react/Icons/MyIcon'
<Icon icon={MyIcon} />
```