feat: Add type prop to SquareAppIcon
This property is based on AppIcon's type property

It is needed to force the type of konnector icons, or else AppIcon will
try to get their icon from the stack first with the wrong app type
This leads to lots of 404 errors in the home application.

Only after that, AppIcon tries to get the icon from the registry and
succeeds.