feat: Forward remaining props to AppIcon
To configure to behavior of AppIcon, in SquareAppIcon, we need access to
all AppIcon properties. SquareAppIcon does not have access to the
context of the app enough to make choices regarding :

Ex :
- priority : stack or registry, to avoid too many 404 requests
in the app

Now SquareAppIcon takes it's own properties and forwards the remaining
ones directly to AppIcon

This way, when the AppIcon changes, we do not need to change
SquareAppIcon.