fix: Correct colors about Alerter buttons
We have an order problem during the build concerning some css classes.
Classes are imported before others, which causes style errors.

In our case, some properties had to be overridden to keep the expected
behavior in the app.

That being said, the Alerter component seems old and not in accordance
with the specifications.