fix(IconButton): Implementation of forwardRef
By wrapping the component when exporting it,
propTypes can't work anymore and we get this kind of warning in console:
```
Warning: forwardRef render functions do not support
propTypes or defaultProps.
Did you accidentally pass a React component?
```

We need to use the forwardRef as implemented on this PR or as here:
https://bit.ly/3ykcWZx

More info: https://github.com/facebook/react/issues/16653