fix: Prevent `AppLinker` to call openApp when target link has same slug
When in Flagship mobile app, `AppLinker` is responsible to call
`openApp` for the specified URL and the specified slug, so the mobile
app can handle navigation to the target app

But if the target slug is the same as the current displayed app, it is
unnecessary to call `openApp` as the mobile app would have to handle
navigation to the same route

Instead we want to use the default `web` behaviour in this scenario.
Which means that the web app would redirect to the target URL by using
classic `window.location` redirection

We only check for the URL's slug. The domain here should not be a
problem because if it is the same, then we stay on the app, and if it
is different, then the mobile app would be responsible to detect it and
to open a dedicated view for the external link

Note: this feature is enabled ONLY IF the `AppLinker` component is
inside a `CozyProvider`. If it is not the case, then `AppLinker` would
always call `openApp` like before