feat: Let the user able to select content
An user reported us that he was not able
to select a text from a PDF (https://github.com/cozy/cozy-drive/issues/2673).

After a quick look, it appears that HammerJS
adds a `user-select:none` inline style in
order to be able to add its gestures things.

We can disable this option when creating the
Hammer object. This is what I've done here
for the desktop version.

I also removed the swipe gesture for desktop
because we can have "conflict" between text
selection and swipe. This is why I didn't
remove this behavior on mobile, since I don't
have the time to test it correctly on several
devices.

It also appears that we don't need the tap/
doubletap gesture anymore in order to display
the controls because the control is now
only displayed on hover. Else wehre, the opacity
is set to false.