feat: Move `useReferencedContactName` call to Qualification component
Before we only showed the owner (in the qualification panel)
on the BottomSheet component(Mobile),
and not on the Drawer component(Desktop)

Now we want to see the owner also on the Drawer component(Desktop)

The BottomSheet no longer has a problem with some data being asynchronous.
I don't know the exact reason, but it has gone through a lot of changes recently.

As a result `useReferencedContactName` has a much better place
in the `Qualification` component and
is thus present on Mobile and Desktop