fix: ImageViewer crashes when render are too fast
The problem has been encountered when updating a page is too fast for example with an useQuery. Each time thee ImageViewer is unmounted, we also teardown all the gestures. The imageViewer crashes because the Gesture object have not yet been defined. This fix checks if Gesture exists before calling any functions on it