feat: Use helper for get random UUID
We recently used the `randomUUID` method of `crypto`,
and added a special configuration for testing with `Jest`.

We should have added a BC at that time, because e2e tests of applications
can fail if they also do not have the correct Jest configuration.

I propose here a better approach,
which will avoid a BC and changes on the application side