chore: Use specific version of node in travis
When installing node, there are sometimes problems downloading the image, which causes the job to error. To remedy this, we force the use of the version contained in the distribution used by Travis (jammy in this case). To do this, you don't have to specify a version in Travis, which means that you use the version in nvmrc