feat: Add isExtraExtraLarge breakpoint
This allows more granularity in the breakpoint for extra large devices between 1200px and 1400px. This change follows the device ranges used by bootstrap. 

BREAKING CHANGE : Before, we used `isExtraLarge` to get the same result now we have to use `isExtraLarge || isExtraExtraLarge`  because isExtraLarge has a limit now and not isExtraExtraLarge