fix(MenuItem): Override `component` prop
We use the `component` prop on the `MenuItem` component
for use our `ListItem` component instead of the MUI one.
If we want to use this same prop on the
`ActionsMenuItem` component
while using the `MenuItem` child component
(to use the `a` tag for example),
we must intercept the `component` prop
and pass it directly to the `ListItem` component.
Thus we do not modify the MUI API of the `ListItem` component.