fix: show spinner when needed in squareappIcon
Previously it was rendered too liberally.
Now it is rendered only when
- Component is in "loading" variant
- Component has an animationState (success or failed),
which triggers after a loading finishes. We need this
because we don't want to hide the spinner as soon as the loading is over
We want to wait for the ending animation too so the transition looks
smooth.

Also refactored the spinner logic out of the main component for clarity