feat: Add images to transpiled folder
When I put InstallFlagshipAppDialog images in /assets/images folder,
they were automatically added in the app bundle even if the app was not
using InstallFlagshipAppDialog.

Now, I put InstallFlagshipAppDialog images in a folder closer to the
InstallFlagshipAppDialog folder, but I configured babel to copy images
in the /transpiled/react folder.

I had to play with babel because babel --copy-files copy every not js files
(so also .md files ...), except with --no-copy-ignored. But to work,
you need to say to babel that it must take .md files into account, and then
that it must ignore them.

See https://github.com/babel/babel/issues/11394#issuecomment-649014113