chore: Bundlemon no longer monitors the `transpiled/react` folder
Following the update of some subdependencies,
the `baseDir` parameter of the bundlemon
configuration no longer works if it is not filled in
or simply with `"."` or `"./"`.

The `transpiled/react` folder now having only
components, it is more relevant that it is on the
application side that the components are monitored by Bundlemon.