fix(ContactsListModal): Load more contacts list
In the case of very many contacts, the list can become very slow,
this trick makes it possible not to load all the contacts
when opening the modal.

It doesn't really solve the problem, since during a search
or if you scroll down, all the contacts end up being displayed
and the modal becomes "laggy" again.