fix(Dialog): Remove usage of RemoveScroll
This suppressed the scrolling of the BottomSheet when it was in a modal. In any case, we no longer need this fix because, in relation to the PR discussion, we no longer use
body-scroll-lock, and the ActionMenu and Overlay components are now deprecated. In addition, we've updated Mui > 4.9.11

There's still the problem in the cozy-ui doc, but it's due to another problem that was identified a long time ago but never resolved.