fix(Theme): Move global background-color and color from `body` to app node
be more responsive to themes, because for the css variable to have the right value, we need to be under a node that carries the CozyTheme class. But in the `html, body` the tags are too high in the DOM, they're not under a CozyTheme tag, so the variable values aren't right.