feat(CozyTheme): Add ignoreItself prop (add `u-dc` css class)
BREAKING CHANGE: `CozyTheme` use now `display: contents` by default so as not to change the style of the page, since it's just a wrapper that supports styles and nothing else. If this causes problems in the structure of your app, you can always revert to the previous state by setting `<CozyTheme ignoreItself={false}>`.