fix(Viewer): Remove WebviewIntentProvider
In commit 97f28ff, WebviewIntentProvider has been added to ViewerContainer, in order to call intent's methods in the container children.

This is not needed because we expect the webapp to
implemented on the root level.

Also, this would produce a bug because
WebviewIntentProvider is not mean to be declared
twice in a same webapp.

When doing so, the child WebviewIntentProvider
would overwrite the parent one on the flagship app side.
So when opening the Viewer, the parent provider
would stop working.