feat: Use expiration date messages from cozy-client
With cozy-client 45.15.0, we mutualized paper metadata formatting in cozy-client because we also use paper metadata formatting in cozy-keys-browser.

But we forgot to mutualize expiration date messages like "Expired" or "Expire in 10 days". It is fixed with cozy-client X.X.X which introduced getExpiredMessage, getExpiresInMessage and getExpirationDescription helpers. We used them here and remove unused translations.