feat: Remove class creation on body and fix layout
On mobile, we've historically had a problem with background color in applications, because the layout doesn't go all the way to the bottom of the screen. So we corrected the problem by adding a background color. Except that for this to be theme-adaptive, you need to have the right class on the element or a parent element. This poses a number of problems: the class flashes while it's being applied (in the app and intents), and the class has to be managed in a complex way if you want to take the variant into account too. So we're going backwards here, not applying a color, but adjusting the layout so that it's fullheight.