fix: We moved unnecessarily `puppeteer` from devDep to dep
puppeteer causes timeout problems for the CI because it downloads Chrome. We had made the package optional, so that it could be ignored in the CI and installed only when necessary. Except that doing so put puppeteer in dep, and therefore spread the problem to the other repo. Here, we permanently remove this package, which must be installed manually locally if necessary, or in the CI.