feat(DatePicker): Add DatePicker component
This component is already in use on the MesPapiers app,
and we're going to need it on cozy-sharing, so we're
going to add it to cozy-ui while making it more generic.

The style has been kept from the original, as it was
a generic need and not specific to the app.

The `@date-io/date-fns` package is blocked in v1
for this reason:
https://github.com/dmtrKovalenko/date-io/releases/tag/v2.0.0