feat: Migrate date formatting to es module
To allow efficient tree shaking by bundlers we need to use only es
module.

Here we migrate the require used to import dynamically the
appropriate date formatting from date-fns. We lose the support of
date formatting in a lot of languages, but we currently support only
en, fr and es languages so :
- it was not used
- it has no senses to display everything in a default language and
just format the date in another language