feat: Add ShortcutDialog
A "dumb" modal that centralizes UI, translations and basic error management.

It can take an io.cozy.files.shortcut object in prop to prefill fields.

On save, it calls the callback with the values entered by the user.

I decided to not build an io.cozy.files.shortcuts object for the output as there are multiple ways to create them. See documentation :
- https://docs.cozy.io/en/cozy-doctypes/docs/io.cozy.files/#shortcuts
- https://docs.cozy.io/en/cozy-stack/shortcuts/#shortcuts